+++ django_typer/management/__init__.py
@@ -1705,6 +1705,12 @@ def _resolve_help(dj_cmd: "TyperCommand"):
                 cmd.help = hlp
             elif not dj_cmd.typer_app.info.help:
                 dj_cmd.typer_app.info.help = hlp
+    elif (
+        dj_cmd.typer_app.info.help
+        and not dj_cmd.is_compound_command
+        and not dj_cmd.typer_app.registered_commands[0].help
+    ):
+        dj_cmd.typer_app.registered_commands[0].help = dj_cmd.typer_app.info.help
 
 
 def _names(tc: t.Union[typer.models.CommandInfo, Typer]) -> t.List[str]: